clean and document code
